requirements.txt
streamlit
groq
huggingface-hub
python-dotenv
Pillow
